# packages/contracts/openapi/aura-note.v1.yaml
openapi: 3.1.0
info:
  title: AURA Note v1 API
  version: 0.1.0
  description: Initial API contract seed for AURA Note standalone and ClinicOS-integrated modes.
servers:
  - url: /api/v1
paths:
...
  /notes/finalized/{noteId}:
    get:
      operationId: getFinalizedNoteReadOnly
      description: CP-1 read-only finalized note placeholder. Does not generate or edit final clinical content.
      parameters:
        - name: noteId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Read-only finalized note placeholder.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalizedNoteResponse'
  /documentation-workspace/appointments/{appointmentId}:
    get:
      operationId: getDocumentationWorkspaceByAppointment
      description: CP-1 implemented Documentation Workspace shell for an appointment-linked note.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Documentation Workspace shell.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentationWorkspaceResponse'
  /documentation-workspace/appointments/{appointmentId}/visit-session/pause:
    post:
      operationId: pauseVisitSession
      description: Pause the CP-1 synthetic visit timer and lock the editor.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '201':
          description: Visit paused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitSessionControlResponse'
  /documentation-workspace/appointments/{appointmentId}/visit-session/resume:
    post:
      operationId: resumeVisitSession
      description: Resume the CP-1 synthetic visit timer and unlock the editor.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '201':
          description: Visit resumed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitSessionControlResponse'
  /documentation-workspace/appointments/{appointmentId}/visit-session/stop:
    post:
      operationId: stopVisitSession
      description: Stop the CP-1 synthetic visit timer and normal recording scaffold.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '201':
          description: Visit stopped.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitSessionControlResponse'
  /documentation-workspace/appointments/{appointmentId}/recording-exception:
    post:
      operationId: approveRecordingException
      description: Approve a no-audio documentation exception without marking normal recording active.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordingExceptionRequest'
      responses:
        '201':
          description: Recording exception approved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitSessionControlResponse'
  /documentation-workspace/appointments/{appointmentId}/recording/permission:
    post:
      operationId: recordMicrophonePermissionState
      description: Record explicit browser microphone permission state as audit-safe metadata only; live audio storage remains disabled.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordMicrophonePermissionRequest'
      responses:
        '201':
          description: Microphone permission state recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingPermissionResponse'
  /documentation-workspace/appointments/{appointmentId}/recording/chunks:
    post:
      operationId: appendRecordingChunkMetadata
      description: Append metadata-only synthetic recording chunk evidence. Raw audio payload persistence is disabled.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppendRecordingChunkRequest'
      responses:
        '201':
          description: Recording chunk metadata accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingChunkResponse'
  /documentation-workspace/appointments/{appointmentId}/recording/retention:
    get:
      operationId: getRecordingRetentionMetadata
      description: Return raw-audio one-week retention metadata and transcript indefinite-retention posture.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Recording retention metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingRetentionResponse'
  /documentation-workspace/appointments/{appointmentId}/transcription/provider-status:
    get:
      operationId: getTranscriptionProviderStatus
      description: Return mock-only transcription provider status and disabled live-provider governance posture.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Transcription provider status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptionProviderStatusResponse'
  /documentation-workspace/appointments/{appointmentId}/transcription/jobs/mock:
    post:
      operationId: processMockTranscriptionJob
      description: Process deterministic mock transcription from metadata-only chunks without calling a live provider.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '201':
          description: Mock transcription job processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptionJobResponse'
...
  /documentation-workspace/appointments/{appointmentId}/transcript:
    get:
      operationId: getMockTranscript
      description: Read CP-1 mock transcript segments. Transcript retention is indefinite.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Mock transcript.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptResponse'
...
  /documentation-workspace/appointments/{appointmentId}/transcript/segments:
    post:
      operationId: appendMockTranscriptSegment
      description: Append a synthetic mock transcript segment while the timer or exception gate allows documentation.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppendTranscriptSegmentRequest'
      responses:
        '201':
          description: Mock transcript segment appended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitSessionControlResponse'
  /documentation-workspace/appointments/{appointmentId}/transcript/segments/{transcriptSegmentId}/correction:
    post:
      operationId: correctTranscriptSegment
      description: Record audit-safe transcript correction history for a mock transcript segment.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
        - name: transcriptSegmentId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectTranscriptSegmentRequest'
      responses:
        '201':
          description: Transcript segment correction recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptCorrectionResponse'
...
  /notes/{noteId}/content:
    get:
      operationId: getNoteContent
      description: Read the API-backed aura_markdown_v1 note document and autosave metadata for a draft or finalized read-only note.
      parameters:
        - name: noteId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Note content loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteContentResponse'
  /notes/{noteId}/content/autosave:
    post:
      operationId: autosaveNoteContent
      description: Autosave timer-gated clinician note content in aura_markdown_v1 and create an audit-safe version snapshot.
      parameters:
        - name: noteId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteContentRequest'
      responses:
        '201':
          description: Note content autosaved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteContentResponse'
  /notes/{noteId}/versions:
    get:
      operationId: listNoteVersions
      description: List audit-safe note content versions for the API-backed editor.
      parameters:
        - name: noteId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Note version history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteVersionsView'
  /notes/{noteId}/versions/{versionId}/restore:
    post:
      operationId: restoreNoteVersion
      description: Restore an earlier note version through the same timer-gated clinician edit boundary.
      parameters:
        - name: noteId
          in: path
          required: true
          schema: { type: string }
        - name: versionId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreNoteVersionRequest'
      responses:
        '201':
          description: Note version restored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteContentResponse'
...
  /notes/{noteId}/suggestions/{suggestionId}/accept:
    post:
      operationId: acceptSuggestion
      description: Accept a draft-only suggestion into Visit Selections. Low-confidence diagnosis candidates require override metadata.
      parameters:
        - name: noteId
          in: path
          required: true
          schema: { type: string }
        - name: suggestionId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuggestionDecisionRequest'
      responses:
        '201':
          description: Suggestion accepted into Visit Selections.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewActionResponse'
...
components:
  schemas:
...
    RestoreNoteVersionRequest:
      type: object
      required: [restoreReason]
      properties:
        restoreReason: { type: string }
...
    FinalizedNoteResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/FinalizedNoteDetail'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    DocumentationWorkspaceResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/DocumentationWorkspace'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    VisitSessionControlResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [appointment, note, visitSession, auditEvent, domainEvents]
          properties:
            appointment:
              $ref: '#/components/schemas/Appointment'
            note:
              $ref: '#/components/schemas/Note'
            visitSession:
              $ref: '#/components/schemas/VisitSession'
            rawAudioRetention:
              $ref: '#/components/schemas/RawAudioRetentionMetadata'
            transcript:
              $ref: '#/components/schemas/TranscriptView'
            auditEvent:
              $ref: '#/components/schemas/AuditEvent'
            domainEvents:
              type: array
              items:
                $ref: '#/components/schemas/AuraNoteEvent'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...
    RecordingExceptionRequest:
      type: object
      required: [exceptionReason]
      properties:
        exceptionReason: { type: string }
...